Extract the JSON MCP codec into cli/integrations/
Move the shared JSON MCP config codec out of the setup command module and
into its own cli/integrations module, alongside the other surface handlers.
_configure_json_mcp and its two thin wrappers, _configure_mcp (.mcp.json)
and _configure_cursor_for_repo (.cursor/mcp.json), are a self-contained
load/parse/mutate/write unit with no dependency on the setup command
surface, so setup.py should not host them. This is a verbatim move with no
behavior change; setup.py now imports the two wrappers it calls directly.
