+++ packages/nauro/tests/test_setup.py
@@ -9,8 +9,8 @@ from typer.testing import CliRunner
 
 from nauro.cli.commands.setup import (
     _configure_codex,
-    _configure_mcp,
 )
+from nauro.cli.integrations.json_mcp import _configure_mcp
 from nauro.cli.integrations.legacy import CLAUDE_MD_END, CLAUDE_MD_START
 from nauro.cli.main import app
 from nauro.store.registry import register_project, register_project_v2
