+++ packages/nauro/tests/test_setup_atomic_writes.py
@@ -31,11 +31,11 @@ from nauro.cli.commands.setup import (
     HOOK_EVENT_NAME,
     HOOK_SUBCOMMAND,
     HOOK_TIMEOUT_SECONDS,
-    _configure_mcp,
     materialize_hooks_claude_code,
     materialize_hooks_codex,
 )
 from nauro.cli.integrations.claude_user_config import _prune_redundant_user_scope_mcp
+from nauro.cli.integrations.json_mcp import _configure_mcp
 from nauro.cli.nauro_command import _find_nauro_command
 from nauro.store import _atomic
